socket file moved to api folder from lib
